backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["'Petar' <'petar@elab.rs'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.0.0"
algokit-utils = "^1.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "*"}
pip-audit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

